+++ src/FontManager.cpp
@@ -39,8 +39,8 @@ TTF_Font* FontManager::UseFont(const std::string& fontName) {
 void FontManager::LoadFont(const std::string& fontName) {
   std::string fontFilePath = fmt::format(fmt::runtime("{}{}"), CoffeeMaker::Utilities::AssetsDirectory(),
                                          fmt::format(fmt::runtime("/fonts/{}.ttf"), fontName));
-  // TODO recognize DPI Scaling
-  TTF_Font* font = TTF_OpenFont(fontFilePath.c_str(), CoffeeMaker::GlobalWindow::Instance()->DPIScale() * 18);
+  TTF_Font* font =
+      TTF_OpenFont(fontFilePath.c_str(), static_cast<int>(CoffeeMaker::GlobalWindow::Instance()->DPIScale() * 18));
   if (font == nullptr) {
     CM_LOGGER_CRITICAL("Could not load font from given filepath: {}", fontFilePath);
     std::string message =
